Fixed SwiftLint warnings in HomeViewModelTests
